stratified-eviction/requirements.txt
transformers==4.57.1
kvpress==0.5.2
numpy==1.26.4
datasets
Pillow
accelerate